move StaticCall to own category
