+++ tests/Rules/UnusedPublicClassMethodRule/TemplatePathsTest.php
@@ -12,7 +12,7 @@ use PHPUnit\Framework\Attributes\DataProvider;
 use TomasVotruba\UnusedPublic\Collectors\AttributeCallableCollector;
 use TomasVotruba\UnusedPublic\Collectors\MethodCallCollector;
 use TomasVotruba\UnusedPublic\Collectors\PublicClassMethodCollector;
-use TomasVotruba\UnusedPublic\Collectors\StaticMethodCallCollector;
+use TomasVotruba\UnusedPublic\Collectors\StaticCall\StaticMethodCallCollector;
 use TomasVotruba\UnusedPublic\Rules\UnusedPublicClassMethodRule;
 
 final class TemplatePathsTest extends RuleTestCase
